requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
PyPDF2==3.0.1
python-multipart==0.0.6
pydantic_settings
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
aiofiles==23.2.1
pytest==7.4.3
